[CLONE_*TID]: Make tsk->user_tid and int so that 64-bit arches work. - futex uses int as its atomic word type, we pass in user_tid to the futex routines, so the types must match

+++ arch/i386/kernel/process.c
@@ -625,7 +625,7 @@ int copy_thread(int nr, unsigned long clone_flags, unsigned long esp,
 	 * Does the userspace VM want the TID cleared on mm_release()?
 	 */
 	if (clone_flags & CLONE_CLEARTID)
-		p->user_tid = (long *) childregs->edx;
+		p->user_tid = (int *) childregs->edx;
 	return 0;
 }
